fix: add check-run-again job (test run)
`check-run-again` is a second job in the GitHub
Actions workflow file, which checks for the result
of the first job `update-readme`.

If `update-readme` fails, `check-run-again` will
run, which essentially re-runs this workflow.

This is to cover for the off-chance that the
pseudo RNG in the `refresh-featured-artist.js`
scripts returns the same index on consecutive run,
which would previously return a run error
due to the tracked `README.md` file having no
changes to commit.

For testing purpose: The step to use Node to run
the script has been omitted. This will force the
`README.md` file to be unchanged for the test run,
to check that the newly added job works.